Build annoy from source.
Annoy seems to be the cause of the illegal instruction error.
Ideally we would build the the package on the same CPU model as the one
used to run the service, but we cannot build the package on Cloud Run.
The next best thing is to trying building the package from source.
The only solution I found was to build the package from the github
repository.