♻️ Extract classes AnnoyStrategy and ANNStrategy.
The new class has the responsibility for the ANN data directory path
which is shared by the reader and the writer.

Also, simplify ANNWriter by creating back-end writers on initialization.

Modify reader and writer back-end constructors to accept a file path
instead of a directory path and provide a uniform interface.